% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combn_C}
\alias{combn_C}
\title{combn replacement}
\usage{
combn_C(x)
}
\arguments{
\item{x}{is an integer vector, representing a sequence starting at 1.}
}
\description{
This function is an adaptation of combn from base. currently the options are limited, and implying a value of m = 2.
}
\examples{
x = 1:4
all(combn(x,2)==combn_C(x))
}

