% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifelseR.R
\name{ifelseC}
\alias{ifelseC}
\title{ifelse replacement}
\usage{
ifelseC(test, yes, no)
}
\arguments{
\item{test}{logical vector or an object that can be compressed to a logical vector}

\item{yes}{numerical, character, integer or logical vector for cases where test is TRUE.}

\item{no}{numerical, character, integer or logical vector for cases where test is FALSE.}
}
\description{
This function is a port of ifelse from base. Usually the behaviour should be identical.
}
\details{
The parameters must have the same length, or, in the case of yes or no, length must be 1.
}
\examples{
test = c(TRUE,FALSE,NA); yes = c(1,1,1); no = c(2,2,2)
all(ifelseC(test, yes, no) == ifelse(test, yes, no))
}
